+++ todo-app/src/containers/TodoContainer.jsx
@@ -1,16 +1,16 @@
 import React, { Component } from 'react'
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { getTodos } from '../actions/todos.js';
+import { getTodos, addTodos } from '../actions/todos.js';
+import TodoPage from '../components/TodoPage.jsx';
 
 class TodoContainer extends Component {
 
   //ただのセットアップなので、まだ何も起こらない
-
   render() {
     return (
       <div>
-        
+        <TodoPage />
       </div>
     )
   }
